+++ lib/form-submit-school-delete.php
@@ -9,9 +9,8 @@
      *   - id
      */
     
-    
-    
-    if (isset($_POST['delete']) && $_POST['delete'] == 'Delete School' && isset($_POST['id']))
+    //- Delete the school based on ID and redirect
+    if (isset($_POST['delete']) && $_POST['delete'] == 'Delete School' && isset($_POST['id']) && $_SESSION['user']['permissions'] >= 1)
         deleteSchool($_POST['id']);
     
-    header ('Location: /createschool');
+    header ('Location: /createschool);
